+++ src/sham.h
@@ -33,8 +33,8 @@
 
 
 void delay_ms(int delay);
-uint8_t sd_write(const char *f, int f_len, const char *t, int t_len);
-char *sd_load(const char *f, int f_len);
+uint8_t sd_write(const char *f, uint16_t f_len, const char *t, uint16_t t_len);
+char *sd_load(const char *f, uint16_t f_len);
 uint8_t sd_list(void);
 uint8_t sd_erase(void);
 void touch_button_parameters(uint16_t timeout, uint16_t threshold);
